requirements.txt
aiomysql==0.2.0
alembic==1.12.0
annotated-types==0.5.0
anyio==3.7.1
certifi==2023.7.22
click==8.1.7
exceptiongroup==1.1.3
fastapi==0.103.1
h11==0.14.0
httpcore==0.18.0
httpx==0.25.0
idna==3.4
iniconfig==2.0.0
Mako==1.2.4
MarkupSafe==2.1.3
mypy==1.5.1
mypy-extensions==1.0.0
packaging==23.1
pluggy==1.3.0
pydantic==2.3.0
pydantic_core==2.6.3
PyMySQL==1.1.0
pytest==7.4.2
python-dotenv==1.0.0
sniffio==1.3.0
SQLAlchemy==2.0.21
starlette==0.27.0
tomli==2.0.1
typing_extensions==4.8.0
uvicorn==0.23.2
